separate i2c expander conf from button expander, add midi expander conf
