+++ firmware/src/medium/conf/button_expander_conf.hh
@@ -1,32 +1,11 @@
 #pragma once
 #include "drivers/gpio_expander.hh"
+#include "i2c_aux_expander_conf.hh"
 
-namespace MetaModule::ButtonExpander
+namespace MetaModule::Expander::Button
 {
 
 constexpr uint8_t MaxAddresses = 8; // Number of addresses the chip supports
-
-// I2C for Aux header (not audio expander header)
-const mdrivlib::I2CConfig i2c_conf = {
-	.I2Cx = I2C2,
-	.SCL = {mdrivlib::GPIO::B, mdrivlib::PinNum::_10, mdrivlib::PinAF::AltFunc4},
-	.SDA = {mdrivlib::GPIO::B, mdrivlib::PinNum::_11, mdrivlib::PinAF::AltFunc4},
-	.timing =
-		{
-			// .PRESC = 0x00,
-			// .SCLDEL_SDADEL = 0x71,
-			// .SCLH = 0x40,
-			// .SCLL = 0x80,
-			.PRESC = 0x40,
-			.SCLDEL_SDADEL = 0x40,
-			.SCLH = 0x0A,
-			.SCLL = 0x1D,
-		},
-	.priority1 = 2,
-	.priority2 = 1,
-	.analog_filter = true,
-	.digital_filter = mdrivlib::I2CConfig::DigitalFilterLevel::_8,
-};
 
 constexpr mdrivlib::GPIO_expander_conf gpio_chip_conf = {
 	.addr = 0b010'0000, // base address. Actaully address(es) will be detected at runtime
@@ -62,4 +41,4 @@ inline uint32_t order_buttons(uint32_t raw_reading) {
 	return buttons;
 }
 
-} // namespace MetaModule::ButtonExpander
+} // namespace MetaModule::Expander::Button
